Add files via upload
OJAYITSOLUTIONS executable file for installation
The Application Interface is divided into 2 major sections. 
1.	Data Manipulation
2.	Dashboard Data Analysis 
Data Manipulation -This section gives the medical officer the room to submit, update ,delete, clear entry, search record, print and view patients records. This section also has a link to the Dashboard Data Analysis section.
Dashboard Data Analysis – Represents the data graph analysis of the system its divided into 4 categories. This section shows the percentages of the rate of infection amongst students living on and off campus. The Dashboard generates a table analysis of the data with the help of a button click. Users are allowed to print a copy of the dashboard.
a. Percentages of students infected with malaria ON and Off Campus base on Gender
b. Percentages of Monthly report of the infected rate.
c. Total Percentages of Year (2020 and 2021) report of the infected rate.
Malaria Alert Signal 
When the rate of   infection is above metrics a warning and sound alert signal is thrown and records are no longer captured in the database. The medical officer would receive a warning text message on his or her mobile phone alerting users on the increase of the infection rate. These signals are used to inform the medical officers of the high rate of infection amongst students. A condition is set to capture the increase rate of infection and the warning alert signal is fixed to the entry of 80 (item>=80).